feat(component): move Search into its own exposed component
