+++ packages/big-design/src/components/Search/index.ts
@@ -1,0 +1,4 @@
+import { SearchProps as _SearchProps } from './types';
+
+export { Search } from './Search';
+export type SearchProps = _SearchProps;
