+++ packages/big-design/src/components/Search/types.ts
@@ -1,0 +1,8 @@
+import { FormProps } from '../Form';
+import { InputProps } from '../Input';
+
+export interface SearchProps {
+  value: InputProps['value'];
+  onChange: InputProps['onChange'];
+  onSubmit: FormProps['onSubmit'];
+}
