--- packages/big-design/src/components/Table/Search/index.ts
@@ -1,1 +1,0 @@
-export * from './Search';
