+++ packages/big-design/src/components/Table/Table.tsx
@@ -13,7 +13,6 @@ import { Head } from './Head';
 import { HeaderCell } from './HeaderCell';
 import { DragIconHeaderCell, HeaderCheckboxCell } from './HeaderCell/HeaderCell';
 import { Row } from './Row';
-import { Search } from './Search';
 import { StyledTable, StyledTableFigure } from './styled';
 import { TableColumn, TableItem, TableProps } from './types';
 
@@ -35,7 +34,6 @@ const InternalTable = <T extends TableItem>(props: TableProps<T>): React.ReactEl
     sortable,
     stickyHeader,
     style,
-    search,
     ...rest
   } = props;
 
@@ -217,20 +215,8 @@ const InternalTable = <T extends TableItem>(props: TableProps<T>): React.ReactEl
     }
 
     return (
-      <Box marginBottom={shouldRenderActions() && !search ? 'none' : 'medium'}>
-        <PillTabs {...filters} />
-      </Box>
-    );
-  };
-
-  const renderSearch = () => {
-    if (!search) {
-      return;
-    }
-
-    return (
       <Box marginBottom={shouldRenderActions() ? 'none' : 'medium'}>
-        <Search {...search} />
+        <PillTabs {...filters} />
       </Box>
     );
   };
@@ -238,7 +224,6 @@ const InternalTable = <T extends TableItem>(props: TableProps<T>): React.ReactEl
   return (
     <>
       {renderPillTabs()}
-      {renderSearch()}
       {shouldRenderActions() && (
         <Actions
           customActions={actions}
