+++ packages/big-design/src/components/Table/spec.tsx
@@ -6,7 +6,6 @@ import { fireEvent, render, screen, waitForElement } from '@test/utils';
 import { PillTabsProps } from '../PillTabs';
 
 import { Table, TableFigure } from './Table';
-import { TableSearch } from './types';
 
 interface SimpleTableOptions {
   className?: string;
@@ -19,7 +18,6 @@ interface SimpleTableOptions {
   itemName?: string;
   style?: CSSProperties;
   pillTabs?: PillTabsProps;
-  search?: TableSearch;
 }
 
 const getSimpleTable = ({
@@ -33,7 +31,6 @@ const getSimpleTable = ({
   items,
   pillTabs,
   style,
-  search,
 }: SimpleTableOptions = {}) => (
   <Table
     className={className}
@@ -44,7 +41,6 @@ const getSimpleTable = ({
     emptyComponent={emptyComponent}
     style={style}
     filters={pillTabs}
-    search={search}
     columns={
       columns || [
         { header: 'Sku', render: ({ sku }) => sku },
@@ -600,54 +596,4 @@ test('it executes the given callback', () => {
   fireEvent.click(inStockBtn);
 
   expect(onPillClick).toHaveBeenCalledWith('in_stock');
-});
-
-describe('test search in the Table', () => {
-  test('renders Table with the search prop', () => {
-    const search = {
-      value: 'Product',
-      onChange: jest.fn(),
-      onSubmit: jest.fn(),
-    };
-
-    const { getByText, getByPlaceholderText } = render(getSimpleTable({ search }));
-    const input = getByPlaceholderText('Search') as HTMLInputElement;
-
-    expect(getByText('Search')).toBeInTheDocument();
-    expect(getByPlaceholderText('Search')).toBeInTheDocument();
-    expect(input.value).toBe(search.value);
-  });
-
-  test('call onChange when user change value in the input', () => {
-    const onChange = jest.fn();
-
-    const search = {
-      onChange,
-      value: '',
-      onSubmit: jest.fn(),
-    };
-
-    const { getByPlaceholderText } = render(getSimpleTable({ search }));
-    const input = getByPlaceholderText('Search') as HTMLInputElement;
-
-    fireEvent.change(input, { target: { value: 'Product' } });
-
-    expect(onChange).toHaveBeenCalled();
-  });
-
-  test('call onSubmit when user click to the Search button', () => {
-    const onSubmit = jest.fn();
-
-    const search = {
-      onSubmit,
-      value: '',
-      onChange: jest.fn(),
-    };
-
-    const { getByText } = render(getSimpleTable({ search }));
-
-    fireEvent.click(getByText('Search'));
-
-    expect(onSubmit).toHaveBeenCalled();
-  });
 });
