+++ packages/docs/PropTables/TablePropTable.tsx
@@ -54,15 +54,6 @@ const tableProps: Prop[] = [
     description: 'See Pill Tabs component for details.',
   },
   {
-    name: 'search',
-    types: <NextLink href="#table-search-prop-table">Search</NextLink>,
-    description: (
-      <>
-        See <NextLink href="#table-search-prop-table">below</NextLink> for usage.
-      </>
-    ),
-  },
-  {
     name: 'selectable',
     types: <NextLink href="#table-selectable-prop-table">Selectable</NextLink>,
     description: (
@@ -206,27 +197,6 @@ const tableSortableProps: Prop[] = [
   },
 ];
 
-const tableSearchProps: Prop[] = [
-  {
-    name: 'value',
-    types: 'string',
-    description: 'Value of the search input',
-    required: true,
-  },
-  {
-    name: 'onChange',
-    types: '(event: React.ChangeEvent<HTMLInputElement>) => void',
-    description: 'Native onChange attribute for a HTML input element.',
-    required: true,
-  },
-  {
-    name: 'onSubmit',
-    types: '(event: React.FormEvent<HTMLFormElement>) => void',
-    description: 'Native onSubmit attribute for a HTML form element.',
-    required: true,
-  },
-];
-
 export const TablePropTable: React.FC<PropTableWrapper> = (props) => (
   <PropTable title="Table" propList={tableProps} {...props} />
 );
@@ -241,8 +211,4 @@ export const TableSelectablePropTable: React.FC<PropTableWrapper> = (props) => (
 
 export const TableSortablePropTable: React.FC<PropTableWrapper> = (props) => (
   <PropTable title="Table[Sortable]" propList={tableSortableProps} {...props} />
-);
-
-export const TableSearchPropTable: React.FC<PropTableWrapper> = (props) => (
-  <PropTable title="Table[Search]" propList={tableSearchProps} {...props} />
 );
